Fix for S3 bucket targets under newer duply version
Apparently there was a change in the script at one point to no longer map
target_user and target_pass as S3 credentials. Exporting the AWS environment
variables so the boto library takes advantage of them is the way to go now.
Tried to do this as cleanly as possible within the perpsective backup target
config.